\name{asImRaster}
\alias{asImRaster}


\title{Convert a raster to an im object}
\description{
Conversion between rasters and spatstat's im objects
}
\usage{
	asImRaster(X, ...)
}

\arguments{
	\item{X}{
A RasterLayer}
\item{...}{ additional arguments}
}

\details{
This function is deceased.  use \code{as.im.RasterLayer} in \code{maptools}
}


